+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [2.0.0] - ???
 
+**[Warning] The language file has changed. delete to allow it to reload.**<br>
+
 ### Changed
 
 - Updated Spanish translation file (thanks to Mosquito for translating)
+
+### Fixed
+
+- fixed help message for /playerplot reload and /playerplot update
 
 ## [1.19.1] - 2020-10-05